+++ sync-for-commerce/RELEASES.md
@@ -83,3 +83,13 @@ Based on:
 - [csharp v2.2.0] sync-for-commerce
 ### Releases
 - [NuGet v2.2.0] https://www.nuget.org/packages/Codat.Sync.Commerce/2.2.0 - sync-for-commerce
+
+## 2024-02-14 12:06:54
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce.yaml
+- Speakeasy CLI 1.179.0 (2.257.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v3.0.0] sync-for-commerce
+### Releases
+- [NuGet v3.0.0] https://www.nuget.org/packages/Codat.Sync.Commerce/3.0.0 - sync-for-commerce